Make thread properites API better
We do not always have access to the thread object, for example to name the current thread if on main or if its a thread owned and created by something else.
Setting via a handle is compatible with anything that has the raw handle, and if in a callback of something not exposing thread ownership set via current works.